+++ src/components/TodoList.tsx
@@ -1,5 +1,6 @@
-import classNames from 'classnames';
+import React from 'react';
 import { Todo } from '../types/Todo';
+import { TodoItem } from './TodoItem';
 
 interface Props {
   sortedTodoes: Todo[];
@@ -13,97 +14,27 @@ interface Props {
 
 export const TodoList: React.FC<Props> = ({
   sortedTodoes,
-  onDelete,
   loadingIds,
+  editingId,
+  onDelete,
   todoStatus,
   renameTodo,
   setEditingId,
-  editingId,
 }) => {
   return (
     <section className="todoapp__main" data-cy="TodoList">
-      {sortedTodoes.map(todo => {
-        const isLoading = loadingIds.includes(todo.id);
-        const isEditing = editingId === todo.id;
-
-        return (
-          <div
-            key={todo.id}
-            data-cy="Todo"
-            className={classNames('todo', {
-              completed: todo.completed,
-              'is-editing': isEditing,
-            })}
-          >
-            <label className="todo__status-label">
-              <input
-                aria-label="Mark todo as completed"
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-                checked={todo.completed}
-                onChange={() => todoStatus(todo)}
-              />
-            </label>
-
-            {isEditing ? (
-              <form
-                onSubmit={e => {
-                  e.preventDefault();
-                  const input = e.currentTarget.querySelector('input');
-
-                  if (input) {
-                    renameTodo(todo, input.value);
-                  }
-                }}
-              >
-                <input
-                  autoFocus
-                  data-cy="TodoTitleField"
-                  type="text"
-                  className="todo__title-field"
-                  defaultValue={todo.title}
-                  onBlur={e => renameTodo(todo, e.target.value)}
-                  onKeyUp={e => {
-                    if (e.key === 'Escape') {
-                      setEditingId(null);
-                    }
-                  }}
-                />
-              </form>
-            ) : (
-              <span
-                data-cy="TodoTitle"
-                className="todo__title"
-                onDoubleClick={() => setEditingId(todo.id)}
-              >
-                {todo.title}
-              </span>
-            )}
-
-            {!isEditing && (
-              <button
-                type="button"
-                className="todo__remove"
-                data-cy="TodoDelete"
-                onClick={() => onDelete(todo.id)}
-              >
-                ×
-              </button>
-            )}
-
-            <div
-              data-cy="TodoLoader"
-              className={classNames('modal', 'overlay', {
-                'is-active': isLoading,
-              })}
-            >
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div>
-        );
-      })}
+      {sortedTodoes.map(todo => (
+        <TodoItem
+          key={todo.id}
+          todo={todo}
+          isLoading={loadingIds.includes(todo.id)}
+          isEditing={editingId === todo.id}
+          onDelete={onDelete}
+          todoStatus={todoStatus}
+          renameTodo={renameTodo}
+          setEditingId={setEditingId}
+        />
+      ))}
     </section>
   );
 };
