+++ src/types/SortOrder.ts
@@ -1,0 +1,5 @@
+export enum SortOrder {
+  ALL = 'all',
+  ACTIVE = 'active',
+  COMPLETED = 'completed',
+}
